devtools/client/webconsole/new-console-output/test/components/test_console-api-call.html

<!DOCTYPE HTML>
<html lang="en">
<head>
  <meta charset="utf8">
  <title>Test for ConsoleApiCall component</title>
  <script type="text/javascript" src="chrome://mochikit/content/tests/SimpleTest/SimpleTest.js"></script>
  <script type="application/javascript;version=1.8" src="head.js"></script>
  <!-- Any copyright is dedicated to the Public Domain.
     - http://creativecommons.org/publicdomain/zero/1.0/ -->
</head>
<body>
<p>Test for ConsoleApiCall component</p>

<script type="text/javascript;version=1.8">
window.onload = Task.async(function* () {
  const { prepareMessage } = require("devtools/client/webconsole/new-console-output/utils/messages");
  const { ConsoleApiCall } = require("devtools/client/webconsole/new-console-output/components/message-types/console-api-call");

  const packet = yield getPacket("console.log('foobar', 'test')", "consoleAPICall");
  const message = prepareMessage(packet);
  const rendered = renderComponent(ConsoleApiCall, {message});

  const messageBody = getMessageBody(rendered);
  const consoleStringNodes = getConsoleStringNodes(messageBody);
  is(consoleStringNodes.length, 2, "ConsoleApiCall outputs expected HTML structure");
  is(messageBody.textContent, "foobar test", "ConsoleApiCall outputs expected text");

  for (let i = 0; i < 3; i++) {
    const countPacket = yield getPacket("console.count('bar')", "consoleAPICall");
    const countMessage = prepareMessage(countPacket);
    const countRendered = renderComponent(ConsoleApiCall, {message: countMessage});
    testConsoleCountRenderedElement(countRendered, `bar: ${i + 1}`);
  }

  SimpleTest.finish()
});

function getMessageBody(renderedComponent) {
  const queryPath = "div.message.cm-s-mozilla span span.message-flex-body " +
    "span.message-body.devtools-monospace";
  return renderedComponent.querySelector(queryPath);
}

function getConsoleStringNodes(messageBody) {
  return messageBody.querySelectorAll("span.console-string");
}

function testConsoleCountRenderedElement(renderedComponent, expectedTextContent) {
  info("Testing console.count rendered element");

  const messageBody = getMessageBody(renderedComponent);
  const consoleStringNodes = getConsoleStringNodes(messageBody);

  is(consoleStringNodes.length, 1,
    "console.count rendered element has the expected HTML structure");
  is(messageBody.textContent, expectedTextContent,
    "console.count rendered element has the expected text content");
}
</script>
</body>
</html>
